requirements.txt
Flask==1.1.1
Flask-CKEditor==0.4.3
Flask-WTF==0.14.2
numpy==1.17.0
pandas==0.25.0
Pillow==6.1.0
gunicorn==19.9.0
opencv-python==4.1.1.26
pytesseract
